Revert "Clean up Heap Server Side (CU-0000000) (#122)"
This reverts commit 66385744a0bfdf09554534973c8a11a7ed1d778b.
