+++ src/utils/heapServerSide.ts
@@ -1,5 +1,62 @@
 import axios from "axios";
 
+import config from "../config";
+
+const { HEAP_API_URL, HEAP_APP_ID } = config;
+
+/**
+ * Methods to use for sending events to Heap on the Server Side
+ */
+
+const heapFetch = async (endpoint: string, body: any) => {
+  try {
+    await axios.post(`${HEAP_API_URL}${endpoint}`, body);
+  } catch (e) {
+    console.error("Unable to call Heap API", e);
+  }
+};
+
+/**
+ * Sends heap events from the server to Heap.
+ *
+ * @param identity
+ * @param eventName
+ * @param properties
+ */
+const trackEventToCallInServer = async (
+  identity: string,
+  eventName: string,
+  properties?: { [key: string]: string },
+) => {
+  const body = {
+    app_id: HEAP_APP_ID,
+    identity,
+    event: eventName,
+    properties: properties || {},
+  };
+
+  await heapFetch("/track", body);
+};
+
+/**
+ * Associates properties to a user in heap, called on the server
+ *
+ * @param identity
+ * @param properties
+ */
+const addAccountPropsToCallInServer = async (
+  identity: string,
+  properties: { [key: string]: string },
+) => {
+  const body = {
+    app_id: HEAP_APP_ID,
+    account_id: identity,
+    properties: properties || {},
+  };
+
+  await heapFetch("/add_account_properties", body);
+};
+
 /**
  * Similar to heapTrack method but uses /api/d to send the event server-side event instead
  * instead of client side. This method should be called in the frontend.
@@ -51,4 +108,9 @@ const heapAddAccountPropsServerSide = async (
   }
 };
 
-export { heapAddAccountPropsServerSide, heapTrackServerSide };
+export {
+  addAccountPropsToCallInServer,
+  heapAddAccountPropsServerSide,
+  heapTrackServerSide,
+  trackEventToCallInServer,
+};
