--- src/utils/serverUtils/README.md
@@ -1,2 +1,0 @@
-Server Side specific utils are required since some npm packages can only be
-imported in a NodeJS environment instead of the browser.
