+++ src/utils/serverUtils/index.ts
@@ -2,4 +2,3 @@
 export * from "./deleteGCSObject";
 export * from "./getHasuraClient";
 export * from "./getOrCreateUser";
-export * from "./heapServerSide";
